Moved getPhoneContacts() to AppRepository from DatabaseService.
