+++ Spika/Data/Services/Database/DatabaseService.swift
@@ -25,11 +25,9 @@ enum DatabaseError: Error {
 
 class DatabaseService {
     let coreDataStack: CoreDataStack
-    let phoneNumberParser: PhoneNumberParser
     
-    init(coreDataStack: CoreDataStack, phoneNumberParser: PhoneNumberParser) {
+    init(coreDataStack: CoreDataStack) {
         self.coreDataStack = coreDataStack
-        self.phoneNumberParser = phoneNumberParser
     }
 }
 
@@ -68,44 +66,6 @@ extension DatabaseService {
                 } catch {
                     NSLog("Core Data Error: saving users: \(error.localizedDescription)")
                     promise(.failure(DatabaseError.savingError))
-                }
-            }
-        }
-    }
-    
-    func getPhoneContacts() -> Future<ContactFetchResult, Error> {
-        return Future() { promise in
-            let store = CNContactStore()
-            var contacts = [FetchedContact]()
-            CNContactStore().requestAccess(for: .contacts) { granted, error in
-                if let error = error {
-                    print("failed to request access", error)
-                    promise(.success(ContactFetchResult(error: error)))
-                    return
-                }
-                
-                if granted {
-                    let keys = [CNContactGivenNameKey, CNContactFamilyNameKey, CNContactPhoneNumbersKey]
-                    let request = CNContactFetchRequest(keysToFetch: keys as [CNKeyDescriptor])
-                    DispatchQueue.global(qos: .background).async {
-//                        let timer = ParkBenchTimer()
-                        do {
-                            try store.enumerateContacts(with: request, usingBlock: { (contact, stopPointer) in
-                                let fetchedContacts = self.phoneNumberParser.parse(contact.phoneNumbers.map { $0.value.stringValue })
-                                //self.phoneNumberParser.phoneNumberKit.parse(contact.phoneNumbers.map { $0.value.stringValue })
-//                                    .map {  self.phoneNumberParser.phoneNumberKit.format($0, toType: .e164, withPrefix: true) }
-                                    .map { FetchedContact(firstName: contact.givenName, lastName: contact.familyName, telephone: $0) }
-                                
-                                contacts.append(contentsOf: fetchedContacts)
-                            })
-                        } catch let error {
-                            print("Failed to enumerate contact", error)
-                        }
-//                        print("Contact Pull finished: \(contacts.count), duration: \(timer.stop())")
-                        promise(.success(ContactFetchResult(fetchedContacts: contacts)))
-                    }
-                } else {
-                    promise(.success(ContactFetchResult()))
                 }
             }
         }
